consolidate damage breakdown chart code
